Refactor UI components by renaming files and updating package structure
